devPyside6-test/requirements.txt
PySide6
qrcode
pillow
